+++ cms/src/features/comments/Comments.tsx
@@ -1,0 +1,9 @@
+const Comments = () => {
+  return (
+    <>
+      <h2>Comments</h2>
+    </>
+  );
+};
+
+export default Comments;