+++ cms/src/features/posts/Posts.tsx
@@ -1,0 +1,9 @@
+const Posts = () => {
+  return (
+    <>
+      <h2>Posts</h2>
+    </>
+  );
+};
+
+export default Posts;