+++ cms/src/hooks/usePrevLocation.tsx
@@ -1,12 +1,9 @@
 import React, {
   createContext,
-  Dispatch,
-  RefObject,
   useContext,
   useEffect,
   useMemo,
   useRef,
-  useState,
 } from "react";
 import { useLocation } from "react-router";